zulip/requirements.txt
zulip
pypng
